+++ src/api/authApi.js
@@ -1,0 +1,1 @@
+// 인증 관련 API 엔드포인트
