+++ src/api/chatApi.js
@@ -1,0 +1,1 @@
+// 채팅 관련 API 엔드포인트
