+++ src/api/sharingApi.js
@@ -1,0 +1,1 @@
+// 나눔하기 관련 API 엔드포인트
